Fix generator circular deps
