+++ packages/generator/src/edmx-to-vdm/edmx-to-vdm-util.ts
@@ -9,12 +9,6 @@ import {
 import { VdmComplexType, VdmEnumType, VdmMappedEdmType } from '../vdm-types';
 import { EdmxAction, EdmxFunction } from '../edmx-parser/v4/edm-types';
 import { EdmxFunctionImportV2 } from '../edmx-parser/v2/edm-types';
-import {
-  complexTypeForName,
-  enumTypeForName,
-  findComplexType,
-  findEnumType
-} from './common/entity';
 
 const logger = createLogger({
   package: 'generator',
@@ -28,24 +22,28 @@ export function stripNamespace(name: string): string {
   const nameParts = name.split('.');
   return nameParts[nameParts.length - 1];
 }
+
 /**
  * @internal
  */
 export function isCollectionType(typeName: string): boolean {
   return collectionRegExp.test(typeName);
 }
+
 /**
  * @internal
  */
 export function isEdmType(typeName: string): boolean {
   return typeName.startsWith('Edm');
 }
+
 /**
  * @internal
  */
 export function complexTypeName(type: string): string | undefined {
   return last(type.split('.'));
 }
+
 /**
  * @internal
  */
@@ -59,6 +57,7 @@ export function parseTypeName(typeName: string): string {
     ? parseCollectionTypeName(typeName)
     : typeName;
 }
+
 /**
  * @internal
  */
@@ -69,12 +68,14 @@ export function parseCollectionTypeName(typeName: string): string {
   }
   return name;
 }
+
 /**
  * @internal
  */
 export function isV2Metadata(metadata: EdmxMetadata): boolean {
   return metadata.oDataVersion === 'v2';
 }
+
 /**
  * @internal
  */
@@ -82,6 +83,7 @@ export function isComplexTypeOrEnumType(typeName: string): boolean {
   const typeParts = typeName.split('.');
   return typeParts[0] !== 'Edm' && typeParts[1] !== undefined;
 }
+
 /**
  * @internal
  */
@@ -93,6 +95,7 @@ export function isComplexType(
     ? !!findComplexType(name, complexTypes)
     : false;
 }
+
 /**
  * @internal
  */
@@ -101,6 +104,7 @@ export function isEnumType(name: string, enumTypes: VdmEnumType[]): boolean {
     ? !!findEnumType(name, enumTypes)
     : false;
 }
+
 /**
  * @internal
  */
@@ -111,12 +115,14 @@ export function checkCollectionKind(property: EdmxProperty): void {
     );
   }
 }
+
 /**
  * @internal
  */
 export function complexTypeFieldType(typeName: string): string {
   return typeName + 'Field';
 }
+
 /**
  * @internal
  */
@@ -135,6 +141,7 @@ export function getTypeMappingActionFunction(
     `Could not get a action/function parameter. '${typeName}' is not an EDM type.`
   );
 }
+
 /**
  * @internal
  */
@@ -175,11 +182,13 @@ export function typesForCollection(
     'Types in inside a collection must be either have complex or EDM types.'
   );
 }
+
 /**
  * @internal
  */
 export const propertyJsType = (type: string): string | undefined =>
   type.startsWith('Edm.') ? edmToTsType(type) : undefined;
+
 /**
  * @internal
  */
@@ -203,4 +212,75 @@ export function hasUnsupportedParameterTypes(
     return true;
   }
   return false;
+}
+
+const getPostfix = (type: string) => last(type.split('.'));
+
+/**
+ * @internal
+ */
+export const findComplexType = (
+  name: string,
+  complexTypes: VdmComplexType[]
+): VdmComplexType | undefined =>
+  complexTypes.find(c => c.originalName === getPostfix(name));
+
+/**
+ * @internal
+ */
+export const findEnumType = (
+  name: string,
+  enumTypes: VdmEnumType[]
+): VdmEnumType | undefined =>
+  enumTypes.find(e => e.originalName === getPostfix(name));
+
+/**
+ * @internal
+ */
+export function complexTypeForName(
+  name: string,
+  complexTypes: VdmComplexType[]
+): string {
+  const complexType = findComplexType(name, complexTypes);
+  if (complexType) {
+    return complexType.typeName;
+  }
+  logger.warn(
+    `No complex type mapping found for ${name}! Using "any" instead. This will most likely result in errors.`
+  );
+  return 'any';
+}
+
+/**
+ * @internal
+ */
+ export function complexTypeFieldForName(
+  name: string,
+  complexTypes: VdmComplexType[]
+): string {
+  const complexType = findComplexType(name, complexTypes);
+  if (complexType) {
+    return complexTypeFieldType(complexType.typeName);
+  }
+  logger.warn(
+    `No complex type mapping found for ${name}! Using "any" instead. This will most likely result in errors.`
+  );
+  return 'any';
+}
+
+/**
+ * @internal
+ */
+export function enumTypeForName(
+  name: string,
+  enumTypes: VdmEnumType[]
+): string {
+  const enumType = findEnumType(name, enumTypes);
+  if (enumType) {
+    return enumType.typeName;
+  }
+  logger.warn(
+    `No enum type mapping found for ${name}! Using "any" instead. This will most likely result in errors.`
+  );
+  return 'any';
 }
